refactor: Moved `STATE_ID_COMPONENT_DELIMITER` to `meltano.core.constants`
