+++ src/meltano/core/constants.py
@@ -1,0 +1,5 @@
+"""Constants used by Meltano."""
+
+from __future__ import annotations
+
+STATE_ID_COMPONENT_DELIMITER = ":"
